Clean up unused adapter code and add clipboard history copy use case (#89)
* refactor(clipboard): rename ClipboardPort to LocalClipboardPort and clean up adapter code

- Renamed ClipboardPort trait to LocalClipboardPort for better clarity
- Removed unused clipboard adapter code from uc-app
- Updated mockall dependency to 0.13.1 for testing
- Simplified clipboard write tests by removing unnecessary clones
- Updated all platform-specific implementations to use new trait name

* 📝 Add docstrings to `mkdir700/tokyo` (#90)

Docstrings generation was requested by @mkdir700.

* https://github.com/UniClipboard/UniClipboard/pull/89#issuecomment-3715026392

The following files were modified:

* `src-tauri/crates/uc-app/src/use_cases/clipboard/copy_from_history_to_system_clipbaord.rs`
* `src-tauri/crates/uc-app/tests/copy_from_history_tests.rs`
* `src-tauri/crates/uc-core/src/clipboard/content.rs`
* `src-tauri/crates/uc-core/src/clipboard/domain.rs`
* `src-tauri/crates/uc-core/src/clipboard/snapshot.rs`
* `src-tauri/crates/uc-core/src/clipboard/tests/content_tests.rs`
* `src-tauri/crates/uc-core/src/clipboard/tests/fixtures.rs`
* `src-tauri/crates/uc-core/src/clipboard/tests/snapshot_tests.rs`
* `src-tauri/crates/uc-core/src/clipboard/tests/view_tests.rs`
* `src-tauri/crates/uc-core/src/clipboard/view.rs`
* `src-tauri/crates/uc-infra/src/db/mapper/clipboard_mapper.rs`
* `src-tauri/crates/uc-infra/src/db/models/clipboard_item_row.rs`
* `src-tauri/crates/uc-infra/src/db/repositories/clipboard_repo.rs`
* `src-tauri/crates/uc-infra/src/fs/blob_store.rs`
* `src-tauri/crates/uc-platform/src/adatpers/clipboard/macos.rs`
* `src-tauri/crates/uc-platform/src/adatpers/clipboard/windows.rs`
* `src-tauri/crates/uc-platform/src/runtime/clipboard/runtime.rs`
* `src-tauri/crates/uc-platform/src/runtime/clipboard/watcher.rs`


* Update src-tauri/crates/uc-infra/src/db/repositories/clipboard_repo.rs


* refactor(clipboard): fix typo in filename from 'clipbaord' to 'clipboard'

- rename file copy_from_history_to_system_clipbaord.rs to copy_from_history_to_system_clipboard.rs
- update module reference in mod.rs to match corrected filename

* refactor(clipboard): clean up unused imports and expose domain types

- Remove unused imports in decision.rs and domain.rs
- Expose ClipboardContentActionDecision and ClipboardContentDecisionDomain in mod.rs
- Clean up redundant permission check function in domain.rs

* Update src-tauri/crates/uc-core/src/clipboard/event.rs


* refactor(clipboard): change size field to optional in clipboard items

- Updated ClipboardItemView struct to make size field optional
- Modified database schema and migrations to allow NULL size values
- Adjusted tests to handle optional size cases
- Updated serialization/deserialization logic for optional size
- Added new test cases for unknown size scenarios

---------

+++ src-tauri/crates/uc-infra/src/db/models/clipboard_item_row.rs
@@ -23,8 +23,8 @@ pub struct ClipboardItemRow {
     /// blob 存储 ID
     pub blob_id: Option<String>,
 
-    /// 内容大小（字节）
-    pub size: Option<i32>,
+    /// 内容大小（字节），NULL 表示大小未知
+    pub size: Option<i64>,
 
     /// MIME 类型
     pub mime: Option<String>,
@@ -39,7 +39,7 @@ pub struct NewClipboardItemRow<'a> {
     pub content_type: &'a str,
     pub content_hash: &'a str,
     pub blob_id: Option<&'a str>,
-    pub size: Option<i32>,
+    pub size: Option<i64>,
     pub mime: Option<&'a str>,
 }
 
@@ -52,14 +52,13 @@ pub struct NewClipboardItemRowOwned {
     pub content_type: String,
     pub content_hash: String,
     pub blob_id: Option<String>,
-    pub size: Option<i32>,
+    pub size: Option<i64>,
     pub mime: Option<String>,
 }
 
 impl<'a> From<&'a NewClipboardItemRowOwned> for NewClipboardItemRow<'a> {
-    /// Creates a `NewClipboardItemRow` that borrows string data from the given owned instance.
+    /// Create a `NewClipboardItemRow` that borrows string data from the given owned instance.
     ///
-    /// The returned `NewClipboardItemRow` contains references to the `owned` instance's string fields;
     /// `blob_id` and `mime` are converted from `Option<String>` to `Option<&str>`.
     ///
     /// # Examples
@@ -75,9 +74,12 @@ impl<'a> From<&'a NewClipboardItemRowOwned> for NewClipboardItemRow<'a> {
     ///     size: Some(42),
     ///     mime: Some("text/plain".to_string()),
     /// };
+    ///
     /// let borrowed: NewClipboardItemRow = (&owned).into();
     /// assert_eq!(borrowed.id, owned.id.as_str());
     /// assert_eq!(borrowed.blob_id, owned.blob_id.as_deref());
+    /// assert_eq!(borrowed.mime, owned.mime.as_deref());
+    /// assert_eq!(borrowed.size, owned.size);
     /// ```
     fn from(owned: &'a NewClipboardItemRowOwned) -> Self {
         NewClipboardItemRow {
